+++ src/main/java/com/sparta/springtrello/common/BaseCode.java
@@ -1,0 +1,5 @@
+package com.sparta.springtrello.common;
+
+public interface BaseCode {
+    public ReasonDto getReasonHttpStatus();
+}
