+++ src/main/java/com/sparta/springtrello/common/exception/InvalidRequestException.java
@@ -1,0 +1,7 @@
+package com.sparta.springtrello.common.exception;
+
+public class InvalidRequestException extends RuntimeException {
+    public InvalidRequestException(String message) {
+        super(message);
+    }
+}
