+++ src/main/java/com/sparta/springtrello/common/exception/ServerException.java
@@ -1,0 +1,8 @@
+package com.sparta.springtrello.common.exception;
+
+public class ServerException extends RuntimeException {
+
+    public ServerException(String message) {
+        super(message);
+    }
+}
