+++ src/main/java/com/sparta/springtrello/domain/member/enums/MemberRole.java
@@ -1,0 +1,5 @@
+package com.sparta.springtrello.domain.member.enums;
+
+public enum MemberRole {
+    WORKSPACE, BOARD, READ_ONLY
+}
