+++ src/main/java/com/sparta/springtrello/domain/member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package com.sparta.springtrello.domain.member.repository;
+
+import com.sparta.springtrello.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
