+++ src/main/java/com/sparta/springtrello/domain/workspace/dto/WorkspaceRequestDto.java
@@ -1,0 +1,16 @@
+package com.sparta.springtrello.domain.workspace.dto;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class WorkspaceRequestDto {
+
+    private String name;
+    private String description;
+
+}
