+++ src/main/java/com/sparta/springtrello/domain/workspace/dto/WorkspaceResponseDto.java
@@ -1,0 +1,15 @@
+package com.sparta.springtrello.domain.workspace.dto;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class WorkspaceResponseDto {
+
+    private Long workspace_id;
+    private String name;
+    private String description;
+
+}
