+++ src/main/java/com/sparta/springtrello/domain/workspace/repository/WorkspaceRepository.java
@@ -1,0 +1,10 @@
+package com.sparta.springtrello.domain.workspace.repository;
+
+import com.sparta.springtrello.domain.workspace.entity.Workspace;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface WorkspaceRepository extends JpaRepository<Workspace, Long> {
+
+
+//    List<Workspace> findAllByUserId(Long userId);
+}
