--- src/main/java/com/sparta/springtrello/member/dto/MemberRequestDto.java
@@ -1,16 +1,0 @@
-package com.sparta.springtrello.member.dto;
-
-
-import com.sparta.springtrello.member.enums.MemberRole;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class MemberRequestDto {
-
-    private MemberRole memberRole;
-
-}
