--- src/main/java/com/sparta/springtrello/member/enums/MemberRole.java
@@ -1,5 +1,0 @@
-package com.sparta.springtrello.member.enums;
-
-public enum MemberRole {
-    WORKSPACE, BOARD, READ_ONLY
-}
