--- src/main/java/com/sparta/springtrello/workspace/dto/WorkspaceRequestDto.java
@@ -1,16 +1,0 @@
-package com.sparta.springtrello.workspace.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class WorkspaceRequestDto {
-
-    private String name;
-    private String description;
-
-}
