--- src/main/java/com/sparta/springtrello/workspace/dto/WorkspaceResponseDto.java
@@ -1,15 +1,0 @@
-package com.sparta.springtrello.workspace.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class WorkspaceResponseDto {
-
-    private Long workspace_id;
-    private String name;
-    private String description;
-
-}
